% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outdet.R
\name{outdet}
\alias{outdet}
\title{Outlier detector}
\usage{
outdet(x, thressd)
}
\arguments{
\item{x}{the numerical vector,}

\item{thressd}{the threshold standard deviation}
}
\value{
the outliers, or 'no outlier detected'
}
\description{
This function takes in a vector and exmaines if there are outliers
}
\examples{
outdet (c(10,4,6,8,9,8,7,6,12,14,11,9,8,4,5,10,14,12,15,7,10,14,24,28), 3)
outdet (c(10,4,6,8,9,8,12,15,7,10,14,24,28), 3)
}
